# laravel/storage/api-docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Control Contratos API",
    "version": "1.0.0",
    "description": "Documentación de la API REST para el sistema de control de contratos."
  },
  "servers": [
    {
      "url": "{{baseUrl}}/api",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "summary": "Iniciar sesión",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Token generado" },
          "401": { "description": "Credenciales incorrectas" }
        }
      }
    },
...
    "/dashboard": {"get":{"summary":"Dashboard resumen","security":[{"sanctum":[]}]}}
  },
  "components": {
    "securitySchemes": {
      "sanctum": {"type":"apiKey","name":"X-XSRF-TOKEN","in":"cookie"}
    }
  }
}